Move to reorganize_data to download_utils.
