+++ src/dan_lab_to_nwb/download_utils/validate_paths.py
