+++ src/dan_lab_to_nwb/huang_2025_tdt/huang_2025_tdt_convert_all_sessions.py
@@ -13,8 +13,8 @@ from pydantic import DirectoryPath, FilePath
 from pymatreader import read_mat
 from tqdm import tqdm
 
+from dan_lab_to_nwb.download_utils.reorganize_data import find_tdt_folders
 from dan_lab_to_nwb.huang_2025_tdt.huang_2025_tdt_convert_session import session_to_nwb
-from dan_lab_to_nwb.huang_2025_tdt.reorganize_data import find_tdt_folders
 
 
 def dataset_to_nwb(
